🎨[Refactor] dateFunction 분리
