+++ tsconfig.node.json
@@ -1,6 +1,6 @@
 {
     "extends": "@electron-toolkit/tsconfig/tsconfig.node.json",
-    "include": ["electron.vite.config.*", "src/preload/**/*", "src/renderer/global.d.ts"],
+    "include": ["electron.vite.config.*", "src/main/**/*", "src/preload/**/*", "src/renderer/global.d.ts"],
     "compilerOptions": {
         "composite": true,
         "types": ["electron-vite/node"],
